+++ core/src/constants.ts
@@ -11,8 +11,6 @@ export const ARENA_OPPONENTS_MAX_GAP = 2;
 export const BRUTE_STARTING_POINTS = 11;
 export const PERKS_TOTAL_ODDS = WEAPONS_TOTAL_ODDS + PETS_TOTAL_ODDS + SKILLS_TOTAL_ODDS;
 
-export const SHIELD_BLOCK_ODDS = 0.45;
-
 export const NO_WEAPON_TOSS = 10;
 
 export const Animations = [
@@ -62,7 +60,7 @@ export const FIGHTER_HIT_ANCHOR = {
   dog: { x: 30, y: 30 },
 } as const;
 
-export const MAX_FAVORITE_BRUTES = 3;
+export const MAX_FAVORITE_BRUTES = 30;
 
 export const BASE_FIGHTER_STATS = {
   reversal: 0,